\name{TigerFlathead}
\alias{TigerFlathead}
\docType{data}
\title{Data for Tiger Flathead in South East Survey}
\description{TigerFlathead is a a data set arising from a south-east Australia fish survey, see Bax and Williams (1999).
}
\usage{TigerFlathead
}
\format{
\describe{
\item{Lon}{The Longitude of the trawl.}
\item{Lat}{The Latitude of the trawl.}
\item{GA_BATHY}{Bathymetry (depth) of the trawl, taken from longitude/latitude of trawls.}
\item{STRING}{The distance along the 100m depth contour. The origin is arbitrary, but is northwards of all trawls.}
\item{AREA_HA}{The swept area of the shot, giving a measure of effort.}
\item{Tiger Flathead_no}{The number of tiger flathead caught.}
\item{Tiger Flathead_wt}{The total biomass of tiger flathead caught.}
}
}
\references{
Bax N, Williams A (eds) (1999) Habitat and fisheries production in the South East Fishery. \emph{Final report to FRDC Project 94/040}, CSIRO Marine Research, Hobart, Tasmania, Australia

Foster, S.D. and Bravington, M.V. (2013) A Poisson-Gamma Model for Analysis of Ecological Non-Negative Continuous Data. \emph{Journal of Environmental and Ecological Statistics} \bold{20}: 533-552
}
\keyword{data}
